pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "localforge"
version = "0.1.0"
description = "Local-first workflow orchestrator — chain local LLMs, image generators, and media tools via YAML recipes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "LocalForge Contributors"},
]
keywords = ["workflow", "orchestrator", "ollama", "local-llm", "automation", "yaml", "pipeline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "PyYAML>=6.0,<7.0",
    "requests>=2.28.0,<3.0",
]

[project.optional-dependencies]
image = ["Pillow>=9.0.0,<11.0"]
full = [
    "Pillow>=9.0.0,<11.0",
    "numpy>=1.24.0,<3.0",
    "rembg>=2.0.0",
]

[project.scripts]
localforge = "localforge.__main__:main"

[project.urls]
Homepage = "https://github.com/gitgoodordietrying/localforge"
Issues = "https://github.com/gitgoodordietrying/localforge/issues"
Repository = "https://github.com/gitgoodordietrying/localforge"

[tool.setuptools.packages.find]
include = ["localforge*", "engine*", "tools*", "clients*", "scripts*"]
